+++ src/globals/plugin-setting.ts
@@ -3,19 +3,26 @@ import { isDevEnvironment } from "src/utils/my-lib";
 
 // exposed to users
 export class OuterSetting {
+	customExtensions: {
+		plaintext: string[];
+	};
+	followObsidianExcludedFiles: boolean;
+	excludedPaths: string[]; // NOTE: can't use Set() or it will be a non-iterable object after deserialization
 	logLevel: LogLevel;
 	enableStopWordsEn: boolean;
 	enableChinesePatch: boolean;
 	enableStopWordsZh: boolean;
 	apiProvider1: ApiProvider;
 	apiProvider2: ApiProvider;
-	excludeExtensions: string[];
 	ui: UISetting;
 }
 
 const isChineseUser = window.localStorage.getItem("language") === "zh";
 
 export const DEFAULT_OUTER_SETTING: OuterSetting = {
+	customExtensions: { plaintext: ["md"] },
+	followObsidianExcludedFiles: true,
+	excludedPaths: [],
 	logLevel: isDevEnvironment ? "trace" : "info",
 	enableStopWordsEn: true,
 	// TODO: 繁体中文
@@ -29,12 +36,11 @@ export const DEFAULT_OUTER_SETTING: OuterSetting = {
 		domain: "",
 		key: "",
 	},
-	excludeExtensions: [],
 	ui: {
 		openInNewPane: true,
 		maxItemResults: 30,
 		showedExtension: "except md",
-		collapseDevSettingByDefault: isDevEnvironment ? false : true
+		collapseDevSettingByDefault: isDevEnvironment ? false : true,
 	},
 };
 
@@ -57,23 +63,21 @@ export type UISetting = {
 // ========== transparent for users ==========
 type InnerSetting = {
 	search: {
-		minTermLengthForPrefix: number;
-		fuzzyProportion: number,
-		minTermLengthForPrefixSearch: number
-		weightFilename: number,
-		weightFolder: number,
-		weightTagText: number,
-		weightHeading: number
-	// weightH1: number;
-	// weightH2: number;
-	// weightH3: number;
-	// weightH4: number;
+		fuzzyProportion: number;
+		minTermLengthForPrefixSearch: number;
+		weightFilename: number;
+		weightFolder: number;
+		weightTagText: number;
+		weightHeading: number;
+		// weightH1: number;
+		// weightH2: number;
+		// weightH3: number;
+		// weightH4: number;
 	};
 };
 
 export const innerSetting: InnerSetting = {
 	search: {
-		minTermLengthForPrefix: 3,
 		fuzzyProportion: 0.2,
 		minTermLengthForPrefixSearch: 2,
 		weightFilename: 3,
@@ -85,5 +89,4 @@ export const innerSetting: InnerSetting = {
 		// weightH3: 1.25,
 		// weightH4: 1.1,
 	},
-
 };