+++ src/services/obsidian/user-data/data-provider.ts
@@ -9,6 +9,7 @@ import {
 	parseFrontMatterAliases,
 	type CachedMetadata,
 } from "obsidian";
+import { OuterSetting } from "src/globals/plugin-setting";
 import type { IndexedDocument } from "src/globals/search-types";
 import { logger } from "src/utils/logger";
 import { TO_BE_IMPL, getInstance } from "src/utils/my-lib";
@@ -21,18 +22,27 @@ import { ViewRegistry, ViewType } from "../view-registry";
 export class DataProvider {
 	private readonly vault = getInstance(Vault);
 	private readonly app = getInstance(App);
-	private readonly supportedExtensions =
-		getInstance(ViewRegistry).supportedExtensions();
+	private readonly setting = getInstance(OuterSetting);
 	private readonly privateApi = getInstance(PrivateApi);
-	private viewRegistry = getInstance(ViewRegistry);
-	private readonly MARKDOWN_LINK_REGEX = /\[([^[\]]+)\]\([^()]*\)/g;
-	private readonly MARKDOWN_ASTERISK_REGEX = /\*\*(.*?)\*\*/g;
-	private readonly MARKDOWN_BACKTICK_REGEX = /`(.*?)`/g;
+	private readonly viewRegistry = getInstance(ViewRegistry);
+	private readonly htmlParser = getInstance(HtmlParser);
+	private excludedPaths: Set<string>;
+	private supportedExtensions:Set<string>;
 	public readonly obsidianFs = this.vault.adapter as FileSystemAdapter;
 
 	private static readonly contentIndexableViewTypes = new Set([
 		ViewType.MARKDOWN,
 	]);
+
+	constructor() {
+		this.init();
+	}
+
+	// update internal states based on OuterSetting
+	init() {
+		this.excludedPaths = new Set(this.setting.excludedPaths)
+		this.supportedExtensions = new Set(this.setting.customExtensions.plaintext);
+	}
 
 	async generateAllIndexedDocuments(
 		files: TFile[],
@@ -96,7 +106,12 @@ export class DataProvider {
 		return (
 			this.supportedExtensions.has(FileUtil.getExtension(path)) &&
 			path.lastIndexOf("excalidraw.md") === -1 &&
-			this.privateApi.isNotExcludedPath(path)
+			(this.setting.followObsidianExcludedFiles
+				? this.privateApi.isNotObsidianExcludedPath(path)
+				: true) &&
+			(this.excludedPaths.size === 0
+				? true
+				: this.isNotCustomExcludedPath(path))
 		);
 	}
 
@@ -116,7 +131,7 @@ export class DataProvider {
 			const plainText = await this.vault.cachedRead(file);
 			// return plainText;
 			return file.extension === "html"
-				? this.extractHtmlContent(plainText)
+				? this.htmlParser.toMarkdown(plainText)
 				: plainText;
 		} else {
 			throw Error(
@@ -147,7 +162,27 @@ export class DataProvider {
 		);
 	}
 
-	private extractHtmlContent(htmlText: string) {
+	private isNotCustomExcludedPath(path: string) {
+		const parts = path.split("/");
+		let currentPath = "";
+
+		for (const part of parts) {
+			currentPath += (currentPath ? "/" : "") + part;
+			if (this.excludedPaths.has(currentPath)) {
+				return false;
+			}
+		}
+		return true;
+	}
+}
+
+@singleton()
+class HtmlParser {
+	private readonly MARKDOWN_LINK_REGEX = /\[([^[\]]+)\]\([^()]*\)/g;
+	private readonly MARKDOWN_ASTERISK_REGEX = /\*\*(.*?)\*\*/g;
+	private readonly MARKDOWN_BACKTICK_REGEX = /`(.*?)`/g;
+
+	toMarkdown(htmlText: string) {
 		// use a replacement function to determine how to replace the matched content
 		let cleanMarkdown = htmlToMarkdown(htmlText);
 		// 使用替换函数来确定如何替换匹配到的内容